refactor: extract type defs to types.ts
